+++ SpaceInvaders_2023/Source/MathTools.h
@@ -4,7 +4,7 @@
 
 	inline float const lineLength(Vector2 A, Vector2 B) noexcept
 	{
-		return sqrtf(pow(B.x - A.x, 2) + pow(B.y - A.y, 2));
+		return sqrtf(static_cast<float>(pow(B.x - A.x, 2) + pow(B.y - A.y, 2)));
 	}
 
 	inline bool const pointInCircle(Vector2 circlePos, float radius, Vector2 point) noexcept
