+++ SpaceInvaders_2023/Source/Walls.cpp
@@ -16,7 +16,7 @@ void Wall::Render(Texture2D texture) const {
         WHITE);
 
     if (health > 0) {
-        DrawText(TextFormat("%i", health), position.x - 10, position.y - 10, 20, RED);
+        DrawText(TextFormat("%i", health), static_cast<int>(position.x) - 10, static_cast<int>(position.y) - 10, 20, RED);
     }
 }
